+++ commons-http-test/src/test/java/com/hb0730/commons/http/test/CommonsHttpClientPoolUtils.java
@@ -3,7 +3,7 @@ package com.hb0730.commons.http.test;
 import com.hb0730.commons.http.CommonHttps;
 import com.hb0730.commons.http.config.HttpConfig;
 import com.hb0730.commons.http.support.httpclient.HttpClientAsyncImpl;
-import com.hb0730.commons.http.utils.HttpAsyncUtils;
+import com.hb0730.commons.http.utils.HttpAsync;
 import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
 import org.apache.hc.client5.http.impl.async.HttpAsyncClients;
 import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManager;
@@ -24,7 +24,7 @@ public class CommonsHttpClientPoolUtils {
 
     private static PoolingAsyncClientConnectionManager connectionManager;
     private static CloseableHttpAsyncClient asyncClient;
-    private static HttpAsyncUtils asyncUtils;
+    private static HttpAsync asyncUtils;
     private final static Object SYNC_LOCK = new Object();
 
     static {
@@ -34,7 +34,7 @@ public class CommonsHttpClientPoolUtils {
                 .build();
     }
 
-    public static HttpAsyncUtils async() {
+    public static HttpAsync async() {
         if (null == asyncClient) {
             synchronized (SYNC_LOCK) {
                 if (null == asyncUtils) {