+++ commons-http-test/src/test/java/com/hb0730/commons/http/test/CommonsHttpClientPoolUtilsTest.java
@@ -1,7 +1,7 @@
 package com.hb0730.commons.http.test;
 
 import com.hb0730.commons.http.support.callback.CommonsNetCall;
-import com.hb0730.commons.http.utils.HttpAsyncUtils;
+import com.hb0730.commons.http.utils.HttpAsync;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -11,7 +11,7 @@ public class CommonsHttpClientPoolUtilsTest {
 
     @Test
     public void asyncTest() throws InterruptedException {
-        HttpAsyncUtils async = CommonsHttpClientPoolUtils.async();
+        HttpAsync async = CommonsHttpClientPoolUtils.async();
         async.get("http://localhost:10000/", new CommonsNetCall() {
             @Override
             public void success(String result) throws IOException {