+++ commons-http-test/src/test/java/com/hb0730/commons/http/test/CommonsOkhttpPoolUtils.java
@@ -3,7 +3,7 @@ package com.hb0730.commons.http.test;
 import com.hb0730.commons.http.CommonHttps;
 import com.hb0730.commons.http.inter.AbstractAsyncHttp;
 import com.hb0730.commons.http.support.okhttp3.OkHttp3AsyncImpl;
-import com.hb0730.commons.http.utils.HttpAsyncUtils;
+import com.hb0730.commons.http.utils.HttpAsync;
 
 /**
  * @author bing_huang
@@ -17,7 +17,7 @@ public class CommonsOkhttpPoolUtils {
         http = new OkHttp3AsyncImpl();
     }
 
-    public static HttpAsyncUtils async() {
+    public static HttpAsync async() {
         return CommonHttps.async().setHttp(http);
     }
 }