+++ commons-http/src/main/java/com/hb0730/commons/http/CommonHttps.java
@@ -1,7 +1,7 @@
 package com.hb0730.commons.http;
 
-import com.hb0730.commons.http.utils.HttpAsyncUtils;
-import com.hb0730.commons.http.utils.HttpSyncUtils;
+import com.hb0730.commons.http.utils.HttpAsync;
+import com.hb0730.commons.http.utils.HttpSync;
 
 /**
  * commons utils list
@@ -10,28 +10,27 @@ import com.hb0730.commons.http.utils.HttpSyncUtils;
  * @since 1.0.0
  */
 public class CommonHttps {
+    public static HttpSync sync = new HttpSync();
+    public static HttpAsync async = new HttpAsync();
 
     /**
-     * 获取同步的utils,每一次调用都会新建，需要自行存储
+     * 获取同步Sync
      *
-     * @return 实例化的 {@link HttpSyncUtils}
-     * @see HttpSyncUtils
+     * @return 实例化的 {@link HttpSync}
+     * @see HttpSync
      */
-    public static HttpSyncUtils sync() {
-        return new HttpSyncUtils();
+    public static HttpSync sync() {
+        return sync;
     }
-
 
     /**
      * 获取支持的异步utils,
      * 每一次调用都会新建，需要自行存储
      *
-     * @return 实例化的 {@link HttpAsyncUtils}
-     * @see HttpAsyncUtils
-     * @see org.apache.hc.core5.http.message.BasicHttpResponse
-     * @see okhttp3.Response
+     * @return 实例化的 {@link HttpAsync}
+     * @see HttpAsync
      */
-    public static HttpAsyncUtils async() {
-        return new HttpAsyncUtils();
+    public static HttpAsync async() {
+        return async;
     }
 }